% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNA.R
\docType{data}
\name{DNA}
\alias{DNA}
\title{DNA Extraction Concentration Data on Coyotes}
\format{The data frame has 30 rows and 2 variables:
\describe{
  \item{kit}{DNA measured with the kit method, in ng/uL}
  \item{chloroform}{DNA measured with the chloroform method, in ng/uL}
  }}
\source{
Riordan, B. (2012). Northeastern Ohio Coyote Hybridization with Wolves. Honors research project, University of Akron, Akron.
}
\usage{
DNA
}
\description{
A total of 30 different coyotes were used for the study. One of the methods was the QIAGEN DNeasy Blood and Tissue Kit and the other was the more traditional chloroform isoamyl alcohol method. Selected randomly, 6 of these coyotes were measured using both methods; 8 were measured with the kit method only and the remaining 16 were measured with chloroform method only.
}
\keyword{datasets}
