% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/race.R
\docType{data}
\name{race}
\alias{race}
\title{5K Race Data}
\format{The data frame has 1051 observations and 2 variables:
\describe{
  \item{Kent}{runner's time completing the Kent course, in seconds}
  \item{Tallmadge}{runner's time completing the Tallmadge course, in seconds}
  }}
\source{
1 \url{https://ritchiessports.com/wp-content/uploads/tallmadge5k/2011-men-overall.txt}

2 \url{https://ritchiessports.com/wp-content/uploads/tallmadge5k/2011-women-overall.txt}

3 \url{https://www.kent.edu/RunTheWorld/past-races}
}
\usage{
race
}
\description{
Runner's time for 2 local 5K races in Northeast Ohio (Kent State University 5K race and Tallmadge Memorial 5K)  were recorded. There were 32 runners who competed in both of these races, 478 who competed in only the Kent race, and 541 who competed in only the Tallmadge race in 2011.
}
\keyword{datasets}
